Separated Start Screen From Negotiation Screen - start game button works
