+++ src/components/Screens/StartScreen.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import "./styles/StartScreen.css";
+import "../styles/StartScreen.css";
 // import bg from '../assets/bg.jpg';
 
 const StartScreen = ({ onStartGame }) => {
@@ -18,8 +18,9 @@ const StartScreen = ({ onStartGame }) => {
   };
   return (
     // <div className="start-screen" style={{ backgroundImage: `url(${bg})` }}>
+    <>
     <div className="start-screen" style={{ backgroundColor: "#f0f0f0" }}>
-      <div className="overlay">
+      <div >
         <h1>Online Used Car Sales Game</h1>
         <section className="description">
           <h2>About the Game</h2>
@@ -53,6 +54,47 @@ const StartScreen = ({ onStartGame }) => {
         </button>
       </div>
     </div>
+      <div style={{ display: "flex", gap: "1rem" }}>
+        <button
+          style={{
+            backgroundColor: "blue",
+            color: "white",
+            boxShadow:
+              "rgba(25, 25, 25, 0.04) 0 0 1px 0, rgba(0, 0, 0, 0.1) 0 3px 4px 0",
+            color: "#008000",
+            cursor: "pointer",
+            display: "inline-block",
+            fontFamily: "Arial, sans-serif",
+            fontSize: "1em",
+            height: "50px",
+            padding: "0 25px",
+            transition: "all 200ms",
+          }}
+          onClick={() => alert("Instructions go here")}
+        >
+          How to play
+        </button>
+        <button
+          style={{
+            backgroundColor: "magenta",
+            color: "white",
+            boxShadow:
+              "rgba(218, 25, 225, 0.04) 0 0 1px 0, rgba(0, 0, 0, 0.1) 0 3px 4px 0",
+            color: "#008000",
+            cursor: "pointer",
+            display: "inline-block",
+            fontFamily: "Arial, sans-serif",
+            fontSize: "1em",
+            height: "50px",
+            padding: "0 25px",
+            transition: "all 200ms",
+          }}
+          onClick={() => alert("Game by Your Name")}
+        >
+          Credits
+        </button>
+      </div>
+      </>
   );
 };
