+++ src/components/styles/GameScreen.css
@@ -3,7 +3,7 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  height: 100vh;
+  height: 100%;
   background-color: #1a1a1a;
   color: white;
 }
