+++ src/components/styles/StartScreen.css
@@ -3,15 +3,15 @@
   background-size: cover;
   background-position: center;
   position: relative;
-  color: white;
+  color: rgb(12, 12, 12);
 }
 
-.overlay {
+/* .overlay {
   background-color: rgba(0, 0, 0, 0.65);
   height: 100%;
   padding: 2rem;
   overflow-y: auto;
-}
+} */
 
 h1 {
   margin-bottom: 1rem;
